Rename test package to service
